+++ src/vue/forms/CreateAssetForm/CreateAssetForm.vue
@@ -8,30 +8,34 @@
       <information-step-form
         v-show="currentStep === STEPS.information.number"
         :is-disabled.sync="isDisabled"
-        @submit="submit"
         @update-is-sellable="setIsSellableAsset"
+        @submit="tryToSubmit"
+        :former="formerForAsset"
       />
 
       <add-quote-assets-step-form
         v-show="currentStep === STEPS.advanced.number"
         :is-disabled.sync="isDisabled"
         @submit="submit"
+        :former="formerForAtomicSwap"
       />
     </form-stepper>
   </div>
 </template>
 
 <script>
-import ManageAssetRequestMixin from './mixins/manage-asset-request.mixin'
-
-import InformationStepForm from './components/information-step-form'
-import AddQuoteAssetsStepForm from './components/add-quote-assets-step-form'
+import InformationStepForm from './components/InformationStepForm'
+import AddQuoteAssetsStepForm from './components/AddQuoteAssetsStepForm'
 import FormStepper from '@/vue/common/FormStepper'
 
 import { Bus } from '@/js/helpers/event-bus'
 import { ErrorHandler } from '@/js/helpers/error-handler'
-
-import { mapGetters } from 'vuex'
+import { AssetFormer } from '@/js/formers/AssetFormer'
+import { AtomicSwapFormer } from '@/js/formers/AtomicSwapFormer'
+import { buildIssuanceCreationOperation } from '@/js/helpers/issuance-creation-helper'
+import { api } from '@/api'
+import config from '@/config'
+import { mapGetters, mapActions } from 'vuex'
 import { vuexTypes } from '@/vuex'
 
 const EVENTS = {
@@ -48,13 +52,23 @@ const STEPS = {
 }
 
 export default {
-  name: 'create-asset-form-simplified-module',
+  name: 'create-asset-form',
   components: {
     InformationStepForm,
     AddQuoteAssetsStepForm,
     FormStepper,
   },
-  mixins: [ManageAssetRequestMixin],
+
+  props: {
+    formerForAsset: {
+      type: AssetFormer,
+      default: () => new AssetFormer(),
+    },
+    formerForAtomicSwap: {
+      type: AtomicSwapFormer,
+      default: () => new AtomicSwapFormer(),
+    },
+  },
 
   data: _ => ({
     isDisabled: false,
@@ -64,11 +78,9 @@ export default {
   }),
 
   computed: {
-    ...mapGetters([
-      vuexTypes.accountId,
-      vuexTypes.businessStatsQuoteAsset,
-    ]),
-
+    ...mapGetters({
+      accountBalanceByCode: vuexTypes.accountBalanceByCode,
+    }),
     getSteps () {
       return {
         information: {
@@ -89,6 +101,10 @@ export default {
   },
 
   methods: {
+    ...mapActions({
+      loadBalances: vuexTypes.LOAD_ACCOUNT_BALANCES_DETAILS,
+    }),
+
     moveToNextStep () {
       this.currentStep++
       if (this.$el.parentElement) {
@@ -96,16 +112,35 @@ export default {
       }
     },
 
-    async submit (form) {
+    tryToSubmit () {
+      return this.isSellableAsset ? this.moveToNextStep() : this.submit()
+    },
+
+    async submit () {
       try {
-        this.collectAssetAttributes(form)
-        if (form.isSellable) {
-          this.moveToNextStep()
-          return
-        }
         this.isDisabled = true
 
-        await this.submitCreateAssetRequest(this.requestId)
+        const operations = await this.formerForAsset.buildOps()
+        await api.postOperations(...operations)
+
+        await this.loadBalances()
+
+        const buildIssuance = await buildIssuanceCreationOperation({
+          assetCode: this.formerForAsset.attrs.assetCode,
+          amount: config.MAX_AMOUNT,
+          receiver:
+            this.accountBalanceByCode(this.formerForAsset.attrs.assetCode).id,
+        })
+        await api.postOperations(buildIssuance)
+
+        if (this.isSellableAsset) {
+          this.formerForAtomicSwap.setAttr('assetCode', this.formerForAsset.attrs.assetCode)
+          this.formerForAtomicSwap.setAttr('price', this.formerForAsset.attrs.price)
+
+          let operation = await this.formerForAtomicSwap.buildOps()
+          await api.postWithSignature('/integrations/marketplace/offers', operation)
+        }
+
         Bus.success('create-asset-form.request-submitted-msg')
         this.$emit(EVENTS.submitted)
       } catch (e) {